backend/tests/requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.11.1
